% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAMR.R
\name{AAMR}
\alias{AAMR}
\title{AAMR: Age-adjusted mortality rate}
\usage{
AAMR(d, pop, Nref, ages = c(), fac = 10000)
}
\arguments{
\item{d}{a vector containing the number of deaths in each age group in the study population.}

\item{pop}{a vector containing the population size for each age group within the study population.}

\item{Nref}{a vector containing the population size for each age group within the reference population.}

\item{ages}{a vector containing the lower limit of each age group.}

\item{fac}{a factor used to represent the mortality rate as "deaths per \code{fac} population". The default is 1e4.}
}
\value{
Returns a list with three components:
\itemize{
  \item \code{crude}: The crude mortality rate.
  \item \code{adj}: The age-adjusted mortality rate.
  \item \code{tabm}: A matrix containing \code{"d"}, \code{"pop"}, \code{"Nref"} and the expected number of deaths at each age group.
}
}
\description{
Calculates the age-adjusted mortality rate (direct standardisation).
}
\examples{
## Example 1
d    <- c(1,14,102,259,381,420,328,297)
pop  <- c(670858,1530547,1591913,1551481,1355325,1068705,604175,332148)
Nref <- c(7058427,15541422,16281290,15382114,12733791,9626735,5432779,2828223)
AAMR(d,pop,Nref,ages=c(15,20,30,40,50,60,70,80,100))

## Example 2
# Bruce et al. (2018), Exercise 3.3.1, p. 111.
d    <- c(15,31,78)
pop  <- c(4100,3000,2900)
Nref <- c(5000,3500,1500)
AAMR(d,pop,Nref,ages=c(65,75,85,100),fac=1000)
}
\references{
Bruce, N., Pope, D., Stanistreet, D. (2018). Quantitative Methods for Health Research: A Practical Interactive Guide to Epidemiology and Statistics. Second Edition. John Wiley & Sons Ltd.
}
