% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpol.R
\name{interpol}
\alias{interpol}
\title{interpol: Geometric interpolation between two years}
\usage{
interpol(pop1, pop2, from = 2010, to = 2022)
}
\arguments{
\item{pop1}{a vector of the initial values.}

\item{pop2}{a vector of the ending values.}

\item{from}{initial year or data point.}

\item{to}{ending year or data point.}
}
\value{
The function returns a list containing two components:
\itemize{
  \item \code{rsx}: The growth rates.
  \item \code{interpdata}: A matrix of the intermediate values.
}
}
\description{
Calculates intermediate values between a starting value and an ending value by finding an average annual growth rate.
}
\details{
This function uses geometric interpolation to calculate intermediate values between a starting value and an ending value. The growth rate by sex (s) and age group (x) is given by
\deqn{r_{sx} = \left(\dfrac{pop_{1sx}}{pop_{2sx}}\right)^{1/\Delta t} - 1}
where \eqn{pop_{1sx}} and \eqn{pop_{2sx}} are the populations for the initial and final data points, respectively, and \eqn{\Delta t} is the time elapsed between the two dates.
}
\examples{
# Example 1
# Laurenti et al. (1985), page 34
# Brazilian population in 1970:  93,215,300
# Brazilian population in 1980: 119,098,992
interpol(93215300, 119098992, 1970, 1980)

# Example 2
pop1 <- c(2126148L, 775746L, 884602L, 957100L, 911673L, 812483L, 747361L, 688740L,
          614103L, 501228L, 386337L, 274949L, 216546L)
pop2 <- c(1787296L, 648467L, 752059L, 783322L, 808350L, 881275L, 892896L, 771218L,
          713233L, 649157L, 581323L, 472760L, 356725L)
out  <- interpol(pop1,pop2,from=2010,to=2022)
out$rsx        # the annual growth rate
out$interpdata # a matrix of the intermediate values
}
\references{
Laurenti, R., Mello Jorge, M.H.P., Lebrão, M.L., Gotlieb, S.L.D. (2005). Estatísticas de Saúde. 2nd edition. São Paulo: EPU. ISBN: 9788512408309.
}
