% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMR.R
\name{SMR}
\alias{SMR}
\title{SMR: Standardized mortality ratio}
\usage{
SMR(d, pop, dref, Nref, ages = c())
}
\arguments{
\item{d}{a vector containing the number of deaths in each age group in the study population.}

\item{pop}{a vector containing the population size for each age group within the study population.}

\item{dref}{a vector containing the number of deaths in each age group in the reference population.}

\item{Nref}{a vector containing the population size for each age group within the reference population.}

\item{ages}{a vector containing the lower limit of each age group.}
}
\value{
Returns a list with six components:
\itemize{
  \item \code{obs}: The observed number of deaths.
  \item \code{exp}: The expected number of death.
  \item \code{smr}: The standardized mortality ratio (SMR).
  \item \code{ci}:  An approximate 95\% confidence interval (CI) for the SMR by using the method proposed by Vandenbroucke (1982).
  \item \code{isr}: The indirectly standardised mortality rate per 10,000 inhabitants, given by 10,000 x SMR x crude death rate for the standard population (see Bruce et al., 2018, p. 110).
  \item \code{tabm}: A matrix containing \code{"d"}, \code{"pop"}, \code{"dref"}, \code{"Nref"} and the expected number of deaths at each age group.
}
}
\description{
Calculates the Standardized Mortality Ratio (SMR), a measure comparing the observed number of deaths in a study population to the expected deaths based on a standard population, adjusted for age and sex.
}
\details{
An SMR is calculated by the indirect method of standardisation. It compares the actual deaths in a study population to the deaths that would be expected if that population had the same age/sex-specific mortality rates as a standard population. SMR is calculated as
\deqn{SMR = \dfrac{observed\ number\ of\ deaths}{expected\ number\ of\ deaths}}
It is commonly used in epidemiology for age-standardized mortality comparisons. An SMR of 1.0 means that the number of observed deaths is equal to the number of expected deaths. An SMR higher than 1.0 indicates higher-than-expected mortality, while an SMR lower than 1.0 indicates lower-than-expected mortality.
}
\examples{
## Example
d    <- c(1,14,102,259,381,420,328,297)
pop  <- c(670858,1530547,1591913,1551481,1355325,1068705,604175,332148)
Nref <- c(7058427,15541422,16281290,15382114,12733791,9626735,5432779,2828223)
dref <- c(2,136,1185,2826,4188,4311,3384,3071)
SMR(d,pop,dref,Nref,ages=c(15,20,30,40,50,60,70,80,100))
}
\references{
Bruce, N., Pope, D., Stanistreet, D. (2018). Quantitative Methods for Health Research: A Practical Interactive Guide to Epidemiology and Statistics. Second Edition. John Wiley & Sons Ltd. ISBN: 9781118665411.

Ulm, K. (1990). Simple method to calculate the confidence interval of a standardized mortality ratio (SMR). American Journal of Epidemiology, 131:373–37, doi:10.1093/oxfordjournals.aje.a115507

Vandenbroucke, J.P. (1982). A shortcut method for calculating the 95 percent confidence interval of the standardized mortality ratio. (Letter). American Journal of Epidemiology, 115:303-4, doi:10.1093/oxfordjournals.aje.a113306
}
